add new component for login
